Adjusting object risks.

+++ sql/tables/object_risks.sql
@@ -11,7 +11,7 @@ CREATE TABLE public.object_risks
     risk_id character varying(5) COLLATE pg_catalog."default" NOT NULL,
     object_id integer NOT NULL,
     premium numeric,
-    premium_for_period integer,
+    premium_for_period numeric,
     is_selected character varying(5) NOT NULL,
     id integer DEFAULT nextval('or_id_seq')
 );
